improve ci with QOL changes & fix emit not returning for prom-tracer (#2)
## Description

Attempt adding more CI to make it easier to check various things as
changes are made.

Would like to automate releases, bumping dependencies and versioning as
well, but figured this would be the first step.

This is practice for CI in general. 

### Also fix emit

The emit signal for prom-tracer wasn't working properly. That's fixed in
this branch as well.

## Changes

- Tune clippy 
- Introduce audit
- Introduce coverage although currently coverage is considered 0%
- group just file
- Introduce structured perf-bench as I've found this helpful for making
changes to the prometheus tracer
- tweak ./set-up.sh script
- introduce this PR template

## How Has This Been Tested?

Ran all commands, saw things were building, hopefully CI passes.

## Fixes

#1

---------

Co-authored-by: Will Marsman <will.marsman@reeldata.ai>